Rename Python package to meteofrance_api
Old Python package name (meteofrance) was for retro compatibility with other
Python package managing weather forecast from Météo-France.
We prefere to have same package and project name.
