+++ docs/reference.rst
@@ -9,12 +9,12 @@ Reference
 meteofrance
 -----------
 
-.. automodule:: meteofrance
+.. automodule:: meteofrance_api
    :members:
 
 
 meteofrance.model
 -----------------
 
-.. automodule:: meteofrance.model
+.. automodule:: meteofrance_api.model
     :members:
