+++ src/meteofrance_api/const.py
