+++ src/meteofrance_api/exceptions.py
