+++ src/meteofrance_api/model/__init__.py
