+++ src/meteofrance_api/model/picture_of_the_day.py
