+++ src/meteofrance_api/model/place.py
