+++ src/meteofrance_api/model/warning.py
