+++ src/meteofrance_api/session.py
