+++ tests/test_picture_of_the_day.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 """Tests Météo-France module. PictureOfTheDay class."""
-from meteofrance import MeteoFranceClient
+from meteofrance_api import MeteoFranceClient
 
 
 def test_picture_of_the_day() -> None:
