+++ tests/test_rain.py
@@ -5,8 +5,8 @@ from unittest.mock import Mock
 import pytest
 import requests
 
-from meteofrance import MeteoFranceClient
-from meteofrance.const import METEOFRANCE_API_URL
+from meteofrance_api import MeteoFranceClient
+from meteofrance_api.const import METEOFRANCE_API_URL
 
 
 def test_rain() -> None:
